+++ src/AdBlocker.cs
@@ -119,6 +119,13 @@ namespace GXLightBrowser
                 }
             }
 
+            if (IsHostOrSubdomain(documentHost, "crunchyroll.com"))
+            {
+                return IsHostOrSubdomain(host, "crunchyroll.com") ||
+                    IsHostOrSubdomain(host, "crunchyrollcdn.com") ||
+                    IsHostOrSubdomain(host, "vrv.co");
+            }
+
             return false;
         }